+++ packages/fighting-design/progress/index.ts
@@ -4,6 +4,7 @@ import { install } from '../_utils'
 
 export const FProgress = install(Progress)
 
+/** progress 组件实例类型 */
 export type ProgressInstance = InstanceType<typeof Progress>
 
 export * from './src/interface.d'