+++ packages/fighting-design/progress/src/interface.d.ts
@@ -1,4 +1,1 @@
 export type { ProgressProps } from './props'
-
-/** 不同类型 */
-export type ProgressType = 'primary' | 'success' | 'danger' | 'warning'